fix: save user data

+++ src/modules/Profile/pages/EditProfile.tsx
@@ -66,6 +66,7 @@ export const EditProfile: React.FC<{
   const [showModalProfile, setShowModalProfile] = useState<boolean>(false);
   const [file, setFile] = useState<any>(null);
   const [imageProfile, setImageProfile] = useState<any>(null);
+  const [deletePhoto, setDeletePhoto] = useState<boolean>(false);
   const [profileImagePath, setProfileImagePath] = useState<any>();
   const [isTester, setIsTester] = useState<boolean>(false);
   const firebaseRef = new FirebaseService();
@@ -168,16 +169,26 @@ export const EditProfile: React.FC<{
       homePhone: newPhone.replace(/[^\d\+]+/g, ''),
     };
 
-    //Deleta a foto antiga do usuário no firabase
-    if (profileImagePath !== null) {
-      await firebaseRef.deleteFS(`${profileImagePath}`);
+    let profileImage = profileImagePath;
+
+    if (imageProfile !== null && profileImagePath !== null) {
+      const isDifferent = imageProfile.includes(profileImagePath.split('/')[3]);
+      //Deleta a foto antiga do usuário no firebase
+      if (!isDifferent) {
+        await firebaseRef.deleteFS(`${profileImagePath}`);
+      }
     }
 
+    if (deletePhoto && profileImagePath !== null) {
+      await firebaseRef.deleteFS(`${profileImagePath}`);
+      profileImage = null;
+    }
     //Salva uma nova foto do usuário no firebase
-    let profileImage;
     if (imageProfile !== null) {
-      profileImage = await firebaseRef.createFS(file);
-      setProfileImagePath(profileImage);
+      if (file !== null) {
+        profileImage = await firebaseRef.createFS(file);
+        setProfileImagePath(profileImage);
+      }
     }
 
     const customField: ProfileCustomFieldsInput[] = [
@@ -452,7 +463,7 @@ export const EditProfile: React.FC<{
                 <Box mt="micro" mb="micro">
                   <TouchableOpacity
                     onPress={() => {
-                      // deleteImageProfile();
+                      setDeletePhoto(true);
                       setFile(null);
                       setImageProfile(null);
                       setShowModalProfile(false);
